Different method for report deletion

+++ tab2html/views.py
@@ -5,22 +5,25 @@ from .models import Report, ReportDirectory
 from .constants import *
 from .utils import *
 
+
 frontend = Blueprint('frontend', __name__)
 
+
 @frontend.route('/')
-def list_subdirs():
+def show_index():
     subdirs = os.walk(report_dir_path()).next()[1]
     return render_template(DIRS_TEMPLATE, subdirs=subdirs)
 
+
 @frontend.route('/<subdir>', methods=['GET', 'POST'])
-def list_reports(subdir):
+def show_subdir(subdir):
     error = None
     report_dir = ReportDirectory(report_dir_path(subdir))
 
+    # TODO: Break up into functions
     if request.method == 'POST':
         action = request.args.get('action', '')
 
-        # TODO: Look into WTForms
         if action == 'upload':
             upload = request.files['file']
             if valid_upload(upload):
@@ -39,7 +42,7 @@ def list_reports(subdir):
                     report_dir.delete_report(slug)
                 except EnvironmentError as e:
                     pass
-            return redirect(url_for('.list_reports', subdir=subdir))
+            return redirect(url_for('.show_subdir', subdir=subdir))
 
     try:
         reports = report_dir.get_reports()
@@ -50,23 +53,25 @@ def list_reports(subdir):
                            error=error)
 
 
-@frontend.route('/<subdir>/<slug>')
+@frontend.route('/<subdir>/<slug>', methods=['GET', 'POST'])
 def show_report(subdir, slug):
     report_dir = ReportDirectory(report_dir_path(subdir))
+
+    # TODO: Break up into functions
+    if request.method == 'POST':
+        action = request.args.get('action', '')
+
+        if action == 'delete':
+            try:
+                report_dir.delete_report(slug)
+            except EnvironmentError as e:
+                raise_errno(e)
+
+            return redirect(url_for('.show_subdir', subdir=subdir))
+
     try:
         report = report_dir.get_report(slug)
     except EnvironmentError as e:
         raise_errno(e)
     
     return render_template(subdir + TEMPLATE_EXT, subdir=subdir, report=report)
-
-
-@frontend.route('/<subdir>/<slug>/delete', methods=['POST'])
-def delete_report(subdir, slug):
-    report_dir = ReportDirectory(report_dir_path(subdir))
-    try:
-        report_dir.delete_report(slug)
-    except EnvironmentError as e:
-        raise_errno(e)
-    
-    return redirect(url_for('.list_reports', subdir=subdir))
